backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pymongo==4.6.1
python-jose[cryptography]==3.3.0
httpx==0.26.0
python-dotenv==1.0.0
pydantic==2.5.3
